Updates now filtered against existing chapters.
